Lecture fichiers (temps et precedences) Apo contraintes (temps et precedences) Apo

+++ header.h
@@ -28,7 +28,7 @@ typedef struct taches* ptaches;
 
 
 void lirefichier_temps (char *  nomFichier, ptaches lesTaches);
-void lirefichier_exclu (char *  nomFichier, ptaches lesTaches, int nbtaches, int operation[nbtaches][2]);
+void lirefichier_exclu (char *  nomFichier, ptaches lesTaches, int nbtaches, float operation[nbtaches][2]);
 void lirefichier_precedent (char *  nomFichier, ptaches lesTaches, int nbprec, int nbtaches);
 int nbTaches ();
 int nbprecedences ();
